戦闘予測HUDの追加＆不具合修正

+++ FreeWarriorMark2/FieldCharacter.h
@@ -220,7 +220,7 @@ public:
 
 	float GetSoldiersPercent()
 	{
-		return m_Soldiers / m_MaxSoldiers;
+		return (float)m_Soldiers / (float)m_MaxSoldiers;
 	}
 
 	bool GetIsDetected()
